AutoMove class as of now
